fix(Readme): automatic bugs & scrollbar display
Readme automatic generation need an overall extra prompt

Scrollbar is designed to avoid the situation that there's too many sections in the readme, which will lead to too many buttons in the ui. But the scrollbar seems to be a bit big now. I tried several method to make them tight but still not tight enough. But now I think it's acceptable.